Backend/requirements.txt
flask-security-too
Flask-Cors
flask-restful
flask 
Flask-SQLAlchemy
setuptools
argon2_cffi
celery>=5.0.0